Binded Return to Confirm and Quit Buttons
Used the bind method to call the functions associated with the confirm
and quit buttons - caused an additional optional parameter in these
functions to account for both pressing Return and clicking with a
mouse.

Bug fix:
Forgot to add the frames of the app to the self object within the class.

Additional:
Enhanced the readability with some of the widgets by spanning the code
for some buttons across several lines.