Moved get_user_data function from main to app
Moved the get_user_data function from main to the data retrieval
app since the get_user_data function acts as a main function to
initialize the app with a tk.Tk() and runs the app's mainloop.
Made a one-sentence note in the doctstring of data_retrieval_app.py
describing the function.