Split up job_search_results into two functions
The job_search_results function in indeed_functions was previously
extracting the urls and opening them in the same function. Hence,
for the benefit of more test coverage and smaller/one job functions,
this function was split in retrieve_job_search_results to extract the
URL from the soup object, and display_job_search_results to open the
URLs in the default browser of the computer.
Updated the main.py script accordingly.