Changed indeed_functions file name to indeed.
Changed according to PEP8 and keeping module names short, and since
the "_functions" was unnecessary, I removed it from the name. Made
the necessary changes in the main file, along with changing the
dictionary name from user_data to data.

Lastly, I noticed that 'import sys' was missing from the data_retrieval_app
file. This was an important import because it allowed for a clean
exit of the app whenever the user clicked the "Quit" button.